# openapi.yaml
openapi: 3.0.0
info:
  title: User Authentication API
  version: 1.0.0
  description: |
    Comprehensive API documentation for the User Authentication Service with Resource-Based Access Control (RBAC).
    
    This API provides:
    - User authentication and authorization
    - Global role-based access control
    - Resource-specific role and permission management
    - OpenID Connect (OIDC) configuration
    - JWT token management
    
    ## Authentication
    Most endpoints require authentication via JWT Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Resource-Based Access Control
    The API supports both global roles (ADMIN, SUPERADMIN) and resource-specific roles.
    - Users can have multiple global roles
    - Users can have different roles for different resources
    - A user can only have one role per resource
    - Permissions are verb-based (read, write, update, delete, etc.)
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
    post:
      tags: [Users]
      summary: Create a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
  /users/v2:
    get:
      tags: [Users]
      summary: Get all users (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
        - name: email
          in: query
          schema:
            type: string
          description: Filter users by email (case-insensitive partial match)
        - name: isActive
          in: query
          schema:
            type: boolean
          description: Filter users by active status (true for active, false for inactive)
      responses:
        '200':
          description: Paginated list of users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
  /users/{id}:
    get:
      tags: [Users]
      summary: Get a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: User not found
    delete:
      tags: [Users]
      summary: Delete a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: User not found
  /users/{id}/super-admin:
    put:
      tags: [Users]
      summary: Update a user's super admin status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserSuperAdminRequest'
      responses:
        '200':
          description: User super admin status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: User not found
  /users/{id}/deactivate:
    put:
      tags: [Users]
      summary: Deactivate a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        '200':
          description: User deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: User not found
  /users/{id}/activate:
    put:
      tags: [Users]
      summary: Activate a user
      description: Activate a deactivated user account. Only SUPERADMIN users can perform this action.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        '204':
          description: User activated successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - only SUPERADMIN can activate users
        '404':
          description: User not found
  /users/assign-role:
    post:
      tags: [Users]
      summary: Assign a resource role to a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignUserResourceRole'
      responses:
        '201':
          description: Role assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResourceRole'
        '400':
          description: Bad request - validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
  /users/revoke-role:
    post:
      tags: [Users]
      summary: Revoke a resource role from a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeUserResourceRole'
      responses:
        '204':
          description: Role revoked successfully
        '400':
          description: Bad request - validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
  /user/update/password:
    post:
      tags: [Users]
      summary: Update user password
      description: |
        Update a user's password. 
        
        **Authorization Rules:**
        - SUPERADMIN users can update any user's password without providing current password
        - Regular users can only update their own password and must provide current password
        
        **Password Requirements:**
        - Minimum 6 characters
        - Must match confirmation password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password updated successfully"
        '400':
          description: Bad request - validation error or invalid current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: User not found
  /user/last-update/creds/{userId}:
    get:
      tags: [Users]
      summary: Get user password update history
      description: |
        Retrieve the password update history for a user.
        
        **Authorization Rules:**
        - SUPERADMIN users can view any user's password update history
        - Regular users can only view their own password update history
        
        **Response:**
        - Returns last update timestamp, who updated it, and total count
        - Returns null values if no password updates have been made
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
          description: The unique identifier of the user
      responses:
        '200':
          description: Password update history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordUpdateHistoryResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: User not found

  # Global RBAC Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization: Bearer {token}"'
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User email address
        isActive:
          type: boolean
          description: Whether the user account is active
        isSuperAdmin:
          type: boolean
          description: Whether the user has super admin privileges
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
          description: Global roles assigned to the user
    UserRole:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/Role'
    CreateUser:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          minLength: 6
          description: User password (minimum 6 characters)
        isSuperAdmin:
          type: boolean
          description: Whether the user should be a super admin
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        nickName:
          type: string
          description: User's nickname
        contactNumber:
          type: string
          description: User's contact number
        reportTo:
          type: string
          format: uuid
          description: ID of the user's manager
...
    UpdateUserSuperAdminRequest:
      type: object
      required:
        - isSuperAdmin
      properties:
        isSuperAdmin:
          type: boolean
          description: Whether the user should be a super admin
    UpdatePasswordRequest:
      type: object
      required:
        - userId
        - userEmail
        - new_password
        - new_password_confirmation
      properties:
        userId:
          type: string
          format: uuid
          description: The unique identifier of the user whose password is being updated
        userEmail:
          type: string
          format: email
          description: The email address of the user whose password is being updated
        current_password:
          type: string
          description: Current password (required for regular users, optional for SUPERADMIN when updating their own password)
        new_password:
          type: string
          minLength: 6
          description: New password (minimum 6 characters)
        new_password_confirmation:
          type: string
          minLength: 6
          description: New password confirmation (must match new_password)
    PasswordUpdateHistoryResponse:
      type: object
      properties:
        last_update:
          type: string
          nullable: true
          description: Date and time of the last password update in MM/DD/YYYY HH:MM:SS format
          example: "10/03/2025 14:22:30"
        updated_by:
          type: string
          nullable: true
          description: User ID of who performed the last password update
          example: "user-uuid"
        how_many:
          type: integer
          description: Total number of password updates for this user
          example: 1
...
    Role:
      type: object
      properties:
        id:
          type: string
          description: Role identifier
        name:
          type: string
          description: Role name
          example: ADMIN
        description:
          type: string
          nullable: true
          description: Role description
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/RolePermission'
          description: Permissions associated with this role
...
    Permission:
      type: object
      properties:
        id:
          type: string
          description: Permission identifier
        name:
          type: string
          description: Permission name
          example: read_users
        description:
          type: string
          nullable: true
          description: Permission description
    RolePermission:
      type: object
      properties:
        permission:
          $ref: '#/components/schemas/Permission'

    # Resource-based RBAC schemas
...
    AssignUserResourceRole:
      type: object
      required:
        - userId
        - roleId
      properties:
        userId:
          type: string
          format: uuid
          description: ID of the user to assign the role to
        roleId:
          type: string
          description: ID of the role to assign
        resourceId:
          type: string
          nullable: true
          description: Optional resource ID for resource-specific roles
    RevokeUserResourceRole:
      type: object
      required:
        - userId
        - roleId
      properties:
        userId:
          type: string
          format: uuid
          description: ID of the user to revoke the role from
        roleId:
          type: string
          description: ID of the role to revoke
        resourceId:
          type: string
          nullable: true
          description: Optional resource ID for resource-specific roles
    UserResourceRole:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: User identifier
        resourceRoleId:
          type: string
          format: uuid
          description: Resource role identifier
        resourceType:
          type: string
          description: Type of the resource
        resourceId:
          type: string
          description: Resource identifier
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages
        hasNext:
          type: boolean
          description: Whether there is a next page
        hasPrev:
          type: boolean
          description: Whether there is a previous page
    PaginatedUsersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
          description: Array of users
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
...
security:
  - bearerAuth: []